2025-week-01/01-llm-pdf-summarizer/requirements.txt
pypdf==4.3.1
sumy==0.11.0
nltk==3.9.1
openai>=1.40.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
